Fix warnings for DebugWriter.
(cherry picked from commit 05a6a1afc9404f64764bd1065d1704073b7d3924)

+++ include/writers/DebugWriter.hpp
@@ -336,7 +336,7 @@ namespace panini
 			::FillConsoleOutputCharacterA(
 				m_output,
 				' ',
-				message.length(),
+				static_cast<DWORD>(message.length()),
 				m_cursor,
 				&written
 			);
@@ -344,7 +344,7 @@ namespace panini
 			::FillConsoleOutputAttribute(
 				m_output,
 				m_screenInfo.wAttributes,
-				message.length(),
+				static_cast<DWORD>(message.length()),
 				m_cursor,
 				&written
 			);
